+++ lib/pages/history/widgets/history_total_information_widget.dart
@@ -13,8 +13,8 @@ class HistoryTotalInformationWidget extends StatelessWidget {
     required this.totalComponent,
   });
 
-  final int totalProfit;
-  final int totalCapital;
+  final double totalProfit;
+  final double totalCapital;
   final int workAmount;
   final int totalComponent;
